fixed mobile view bug
